fix: sending references selected auth
